+++ internal/shared/shared.go
